backend/requirements.txt
# Early Intervention System — Python Dependencies
# Install with: pip install -r requirements.txt

# Core Django
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# JWT Authentication
djangorestframework-simplejwt==5.3.0

# WebSockets (real-time alerts)
channels==4.0.0

# PostgreSQL connector (install even if using SQLite now)
psycopg2-binary==2.9.9

# Read .env file
python-decouple==3.8

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
